+++ examples/ParakeetCTC/parakeet_ctc/inference.py
@@ -18,6 +18,7 @@ import os
 from logging import basicConfig, getLogger
 from pathlib import Path
 
+import torch
 from nemo.collections.asr.parts.mixins.transcription import InternalTranscribeConfig, TranscribeConfig
 
 from aitune.torch import load
@@ -44,7 +45,7 @@ def do_inference(
             override_config=TranscribeConfig(
                 batch_size=kwargs["batch_size"],
                 verbose=False,
-                _internal=InternalTranscribeConfig(device="cuda"),
+                _internal=InternalTranscribeConfig(device=torch.device("cuda")),
             ),
             verbose=False,
         )