+++ tests/functional/pytorch/022_aitune_torch_tune_resnet_modelopt_fp16_test.py
@@ -45,7 +45,9 @@ def do_test(backend: TensorRTBackend, dtype: torch.dtype):
         out = model(sample)
     expected_probs = torch.nn.functional.softmax(out[0], dim=0)
 
-    module = Module(model, "functional-resnet18", strategy=OneBackendStrategy(backend))
+    module = Module(
+        model, "functional-resnet18", strategy=OneBackendStrategy(backend).enable_find_max_batch_size(False)
+    )
     # when
     tune(module, data, batch_sizes=[2, 1], dry_run=False, disable_external_logging=False)
     # then - verify tuning